Switch phones if PUK locked in DSDS configuration
There can be cases that the Phone selected for emergency call
is not served in the modem since the modem stack corresponding
to the selected Phone has been stopped due to the SIM PIN/PUK
locked state while there is another stack with an active SIM.

In that case, setting emergency mode in the selected modem can
cause abnormal behavior in the modem side.

Determine whether the modem is active or not before setting
emergency mode, and switch Phones if it's not active.

Bug: 343862428
Test: atest EmergencyStateTrackerTest
Manual test:
Step 1. Power ON the device with two SIMs in 5G HOME only and
        3rd-party LTE coverage with 5G capapable SIM1 and not 5G capable SIM2.
Step 2. Set the SIM1 preferred for calls and data.
Step 3. Go to SIM lock settings and enter incorrect SIM PIN for SIM1, 3 times.
Step 4. Confirm that PUK lock screen is displayed.
Step 5. Initiate an emergency call and observe the behavior.

Change-Id: I8a7e62900c30c35061f1fc63c952b5f7032f1788