Fix PhoneNumberUtilsTest#testFormatSingaporeNational to work with FLAG_NATIONAL_COUNTRY_CODE_FORMATTING_FOR_LOCAL_CALLS.
PhoneNumberUtils.formatNumber() originally had a flag for Singaporean number to be formatted and then another flag was added for TW numbers which included a refactor that encompassed Singaporean numbers as well. This was causing the test to fail as it was expecting not to format the number, but did as it was possible to do with two different flags.

Flag: com.android.internal.telephony.flags.national_country_code_formatting_for_local_calls

Bug: 349464585
Bug: 349745908
Test: atest PhoneNumberUtilsTest

Change-Id: Ief94e0201617be5c9570158a6369783b93a3aeb2