Fix power consumption problem in cast scene
Issue found at,
When user ends the cast but does not operate the phone anymore,
only the interface onDisplayRemoved will be called back and interface onDisplayChanged will not be called back. But there is no processing of onDisplayRemoved in
com.android.internal.telphony.DeviceStateMonitor#mDisplayListener.

Therefore, in DeviceStateMonitor, the on-off status check of the
screen will not be triggered and modem will continue to report the network status based on the previously setUnsolResponseFilter(filter is 127), eventuallyleading to power consumption problem.

Bug: 348570675
Test: Device test and atest
Flag: EXEMPT bugfix
Change-Id: Id3d62f5c55ddda0772ff0c3e0e841dc6fbb975f7