HSUM Message sent when MainUser unlocks
Currently, when the System unlocks, the system checks whether there are
any pending received messages (in the raw table) and processes them if
there are. On HSUM, this should happen when the MainUser unlocks, not
the System, since the MainUser is the one to handle received messages.

This also ensures that while-locked Message notifications show for the
correct (main) user (although not necessarily that it works).

Flag: com.android.internal.telephony.flags.sms_mms_deliver_broadcasts_redirect_to_main_user
Bug: 354051219
Test: send a message to CE-locked MainUser, then unlock and make sure
      the message shows up (on both HSUM and non-HSUM devices)
Test: atest com.android.internal.telephony.gsm.GsmInboundSmsHandlerTest
Change-Id: I5de0b86576acf48a1d0d4a47c08efb9bcac41cc9