Deliver callbacks also when SIM file is missing.
SIM Records read requests are supposed to execute callback at either
success or failure. The latter happens when there's an exception while
reading it, but not when the file is simply missing.

This change fixes this omission, letting the operation fail fast instead
of waiting for 5 minute CTS timeout. In this specific case, ar.exception
was not null, but ar.userObj was.

Bug: 355008150
Test: atest TelephonyManagerTest
Flag: EXEMPT bugfix
Change-Id: If45113e7d8741be14d4a77b00f0337b165d0ffbd