Remove phone from telephony network request
The new TelephonyNetworkProvider would be per-device
instead of per-phone. On dual SIM device, the telephony network
provider would be the singleton communicating with connectivity
service and two TelephonyNetworkController.

Therefore we need to remove the phone depdency from
TelephonyNetworkRequest at the provider layer. This is the preliminary
change before adding TelephonyNetworkProvider.

Flag: EXEMPT refactor
Bug: 343370895
Test: atest TelephonyNetworkFactoryTest DataNetworkTest
Test: Basic telephony functionality tests
Change-Id: Ifc562b6e17c5b5d8817d95192b9352ad4544bec8