Update Usage Setting After Success Rsp from Modem
After calling the HAL to set the usage setting, the
cached value in the Phone needs to be updated so that
future settings to a value different from the original
value will still be sent to the modem rather than
dropped.

This patch holds the desired value in the response
message parameters and just applies it once the rsp
is successful. This does *not* protect against multiple
rapid calls to the modem with redundant new values,
but that shouldn't happen in practice, and the
consequence would be some unnecessary (but harmless)
redundant calls to the HAL.

Bug: 368068457
Test: atest GsmCdmaPhoneTest#testUsageSettingUpdate_ResetToDefault
Test: manually verified on Shiba
Flag: EXEMPT bugfix
Change-Id: Iccbc9981115137bb7c4f5feb5982d31a5d469cd6